Moved the transformUrl logic to a different file for testing. Added new tests.
