+++ packages/bruno-app/src/utils/collections/export.js
@@ -79,4 +79,76 @@ export const exportCollection = (collection) => {
   FileSaver.saveAs(fileBlob, fileName);
 };
 
+/**
+ * Transforms a given URL string into an object representing the protocol, host, path, query, and variables.
+ *
+ * @param {string} url - The raw URL to be transformed.
+ * @param {Object} params - The params object.
+ * @returns {Object} An object containing the URL's protocol, host, path, query, and variables.
+ */
+export const transformUrl = (url, params) => {
+  const urlRegexPatterns = {
+    protocolAndRestSeparator: /:\/\//,
+    hostAndPathSeparator: /\/(.+)/,
+    domainSegmentSeparator: /\./,
+    pathSegmentSeparator: /\//,
+    queryStringSeparator: /\?/
+  };
+
+  const postmanUrl = { raw: url };
+
+  /**
+   * Splits a URL into its protocol, host and path.
+   *
+   * @param {string} url - The URL to be split.
+   * @returns {Object} An object containing the protocol and the raw host/path string.
+   */
+  const splitUrl = (url) => {
+    const urlParts = url.split(urlRegexPatterns.protocolAndRestSeparator);
+    if (urlParts.length === 1) {
+      return { protocol: '', rawHostAndPath: urlParts[0] };
+    } else if (urlParts.length === 2) {
+      const [hostAndPath, _] = urlParts[1].split(urlRegexPatterns.queryStringSeparator);
+      return { protocol: urlParts[0], rawHostAndPath: hostAndPath };
+    } else {
+      throw new Error(`Invalid URL format: ${url}`);
+    }
+  };
+
+  /**
+   * Splits the host and path from a raw host/path string.
+   *
+   * @param {string} rawHostAndPath - The raw host and path string to be split.
+   * @returns {Object} An object containing the host and path.
+   */
+  const splitHostAndPath = (rawHostAndPath) => {
+    const [host, path = ''] = rawHostAndPath.split(urlRegexPatterns.hostAndPathSeparator);
+    return { host, path };
+  };
+
+  try {
+    const { protocol, rawHostAndPath } = splitUrl(url);
+    postmanUrl.protocol = protocol;
+
+    const { host, path } = splitHostAndPath(rawHostAndPath);
+    postmanUrl.host = host.split(urlRegexPatterns.domainSegmentSeparator);
+    postmanUrl.path = path ? path.split(urlRegexPatterns.pathSegmentSeparator).filter(Boolean) : [];
+  } catch (error) {
+    console.error(error.message);
+    return {};
+  }
+
+  // Construct query params.
+  postmanUrl.query = params
+    .filter((param) => param.type === 'query')
+    .map(({ name, value, description }) => ({ key: name, value, description }));
+  
+  // Construct path params.
+  postmanUrl.variable = params
+    .filter((param) => param.type === 'path')
+    .map(({ name, value, description }) => ({ key: name, value, description }));
+
+  return postmanUrl;
+};
+
 export default exportCollection;
