+++ packages/bruno-app/src/utils/exporters/postman-collection.spec.js
@@ -1,0 +1,62 @@
+const { transformUrl } = require('../collections/export');
+
+describe('transformUrl', () => {
+  it('should handle basic URL with path variables', () => {
+    const url = 'https://example.com/{{username}}/api/resource/:id';
+    const params = [
+      { name: 'id', value: '123', type: 'path' },
+    ];
+
+    const result = transformUrl(url, params);
+
+    expect(result).toEqual({
+      raw: 'https://example.com/{{username}}/api/resource/:id',
+      protocol: 'https',
+      host: ['example', 'com'],
+      path: ['{{username}}', 'api', 'resource', ':id'],
+      query: [],
+      variable: [
+        { key: 'id', value: '123' },
+      ]
+    });
+  });
+
+  it('should handle URL with query parameters', () => {
+    const url = 'https://example.com/api/resource?limit=10&offset=20';
+    const params = [
+      { name: 'limit', value: '10', type: 'query' },
+      { name: 'offset', value: '20', type: 'query' }
+    ];
+
+    const result = transformUrl(url, params);
+
+    expect(result).toEqual({
+      raw: 'https://example.com/api/resource?limit=10&offset=20',
+      protocol: 'https',
+      host: ['example', 'com'],
+      path: ['api', 'resource'],
+      query: [
+        { key: 'limit', value: '10' },
+        { key: 'offset', value: '20' }
+      ],
+      variable: []
+    });
+  });
+
+  it('should handle URL without protocol', () => {
+    const url = 'example.com/api/resource';
+    const params = [];
+
+    const result = transformUrl(url, params);
+
+    expect(result).toEqual({
+      raw: 'example.com/api/resource',
+      protocol: '',
+      host: ['example', 'com'],
+      path: ['api', 'resource'],
+      query: [],
+      variable: []
+    });
+  });
+});
+
